Python/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
sqlalchemy==2.0.30
pydantic==2.7.1
httpx==0.27.0
loguru==0.7.2
python-dotenv==1.0.1
aiosqlite==0.20.0
numpy==1.26.4